trading and production and consumption systems
also reworked the town market info hugely and fixed some UI bugs that occurred while travelling
and organised some scripts into a folder for UI system stuff

+++ scripts/TownData.cs
@@ -7,7 +7,7 @@ public partial class TownData : Resource
     [Export] public string townName;
     [Export] public int population;
     [Export] public int wealth;
-    [Export] public int[] stocks = new int[3]; // quantities of items
-    [Export] public int[] production = new int[3]; // items produced per day?
-    [Export] public int[] consumption = new int[3]; // items consumed per day
+    [Export] public float[] stocks = new float[3]; // quantities of items
+    [Export] public float[] production = new float[3]; // items produced per day
+    [Export] public float[] consumption = new float[3]; // items consumed per day
 }
